chore: Migrate remaining tests to Mocha
This commit updates all the non-Puppeteer unit tests to run using Mocha
and then deletes the custom testrunner framework from this repository.
The documentation has also been updated.
