+++ .eslintignore
@@ -2,9 +2,7 @@ test/assets/modernizr.js
 third_party/*
 utils/browser/puppeteer-web.js
 utils/doclint/check_public_api/test/
-utils/testrunner/examples/
 node6/*
 node6-test/*
-node6-testrunner/*
 experimental/
 lib/
