+++ mocha-config/browser-bundle-tests.js
@@ -1,0 +1,6 @@
+const base = require('./base');
+
+module.exports = {
+  ...base,
+  spec: 'utils/browser/*.spec.js',
+};
