+++ mocha-config/doclint-tests.js
@@ -1,0 +1,6 @@
+const base = require('./base');
+
+module.exports = {
+  ...base,
+  spec: 'utils/doclint/**/*.spec.js',
+};
