+++ mocha-config/puppeteer-unit-tests.js
@@ -1,6 +1,8 @@
+const base = require('./base');
+
 module.exports = {
+  ...base,
   file: ['./test/mocha-utils.js'],
   spec: 'test/*.spec.js',
-  reporter: 'dot',
   timeout: process.env.PUPPETEER_PRODUCT === 'firefox' ? 15 * 1000 : 10 * 1000,
 };
