--- utils/testrunner/examples/unhandledpromiserejection.js
@@ -1,35 +1,0 @@
-/**
- * Copyright 2017 Google Inc. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-const {TestRunner, Reporter} = require('..');
-
-const runner = new TestRunner();
-const reporter = new Reporter(runner);
-
-const {describe, xdescribe, fdescribe} = runner;
-const {it, fit, xit} = runner;
-const {beforeAll, beforeEach, afterAll, afterEach} = runner;
-
-describe('testsuite', () => {
-  it('failure', async (state) => {
-    Promise.reject(new Error('fail!'));
-  });
-  it('slow', async () => {
-    await new Promise(x => setTimeout(x, 1000));
-  });
-});
-
-runner.run();
