--- utils/testrunner/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "@pptr/testrunner",
-  "version": "0.8.0",
-  "description": "Puppeteer testrunner",
-  "main": "index.js",
-  "directories": {
-    "example": "examples"
-  },
-  "scripts": {
-    "test": "node test/test.js"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/puppeteer/puppeteer/tree/master/utils/testrunner"
-  },
-  "author": "The Chromium Authors",
-  "license": "Apache-2.0"
-}
